feat(batch): 增加产品系列字段并优化相关功能
- 在 Batch 实体中添加 productSeries 字段，用于表示产品系列- 在批量导入时，将产品系列信息添加到导入数据中- 优化 BatchServiceImpl 中的批量注册设备方法，提高代码可读性
- 调整 PointNewServiceImpl 中的导出功能，去除不必要的用户信息获取

+++ src/main/java/com/xiliulou/afterserver/web/vo/ProductNewDeliverVo.java
@@ -17,7 +17,7 @@ public class ProductNewDeliverVo {
     /**
      * 产品批次id
      */
-    private Integer batchId;
+    private Long batchId;
     /**
      * 产品批次no
      */
